fix: Move useDarkMode in hooks folder
