+++ src/app/account/PageRegister.tsx
@@ -23,7 +23,7 @@ import { Link as RouterLink } from 'react-router-dom';
 
 import { useCreateAccount } from '@/app/account/account.service';
 import { FieldInput, SlideIn, useToastError } from '@/components';
-import { useDarkMode } from '@/utils/darkMode';
+import { useDarkMode } from '@/hooks/useDarkMode';
 
 export const PageRegister = () => {
   const { colorModeValue } = useDarkMode();
