+++ src/app/layout/Page/docs.stories.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { Flex } from '@chakra-ui/react';
 
 import { Layout } from '@/app/layout';
-import { useDarkMode } from '@/utils/darkMode';
+import { useDarkMode } from '@/hooks/useDarkMode';
 
 import { Page, PageTopBar, PageContent, PageBottomBar } from './index';
