+++ src/components/DataList/docs.stories.tsx
@@ -27,7 +27,7 @@ import {
   PaginationButtonPrevPage,
   PaginationInfo,
 } from '@/components';
-import { useDarkMode } from '@/utils/darkMode';
+import { useDarkMode } from '@/hooks/useDarkMode';
 
 import {
   DataList,
